Do not suppress opacity transition for tooltipped-no-delay
As far as I understand (the git history is not too easy to follow on this repo :/), when `tooltipped-no-delay` was added, there were no animations/transitions on the tooltip.

This commit removes the delay, so that the transition starts immediately, instead of suppressing the transition entirely. It effectively makes the 2 variants differ by only the initial delay (like the name suggests).